refactor: deduplicate CopilotUsage types and copilotUsageFrom* helpers
- Centralize CopilotUsage type as alias of CopilotUsageTokens in token-usage/store
- Add copilotUsageToTokens helper in token-usage/index
- Replace 4 inline CopilotUsage interface definitions across 2 service files and anthropic-types with re-export/import of single source
- Replace 5 copilotUsageFromResponse/responsesCopilotUsageFromResponse helper definitions across 4 route handlers with shared copilotUsageToTokens
